requirement.txt
python-telegram-bot==20.3
pytesseract
Pillow
openai
google-api-python-client
google-auth